A u32 count becomes an index through one total conversion
`cadmpeg_core::decode::index_from_u32` widens a u32 count, identifier or index
to a usize index. It is preceded by `const _: () = assert!(usize::BITS >= 32,
...)`: the cast is total on every target the assertion admits, so the
conversion states no refusal. `git grep -n 'unwrap_or(usize::MAX)' -- crates`
is empty.

The brief's census said 41 production sites spell
`usize::try_from(<value>).unwrap_or(usize::MAX)`. The code says otherwise for
three of them: `catia/solve/mesh_quotient.rs:3595` is `Iterator::min` on a
non-empty vector, `:7300` is `checked_shl`, and `inventor/sketch.rs:1806` is
`Iterator::min` over a filtered lookup. Those three are closed by their own
shape, below.

u32 operands, now `index_from_u32` (30 sites, every operand read at the site):
  creo/curve.rs:963, creo/decode/sketch/radii.rs:167 (`radius_ref: u32`,
  definitions.rs:414), creo/feature/definitions.rs 18 sites (every count is a
  `psb::compact_int` result, `psb.rs:178 -> (u32, usize)`), creo/feature/rows.rs:851,
  creo/scalar.rs:206, f3d/design/decode/body.rs:55 (`View::u32_le`),
  f3d/design/decode/operands.rs:2111,2144 (`View::u32_le_at`),
  f3d/writer/generate/records.rs:710 (`degree: u32`),
  iges/entities/geometry.rs:1867 (`degree: u32`).

u64 operands, one shape each, no sentinel:

- creo/decode/build/coverage.rs:910 - `NumericPayload::element_count` returned
  u64. `NumericArray::try_new` already refuses an array whose extents overflow;
  the fold now runs in usize, so an array with more elements than the address
  space can index is refused there and `element_count` returns an index. That
  is the refusal shape, moved to the one constructor that can state it.
- rhino decode.rs:179,180,1734 and mesh.rs:127,160 - all five took the lower of
  a u64 session limit and a usize rhino ceiling. `decode::session_ceiling`
  states it once: a limit the address space cannot name is above every ceiling,
  so the ceiling is the answer. The minimum itself is the pre-existing session
  policy, not a fold.
- container/compound.rs:1508 - `prefix_len` is already an index, so a
  `max_bytes` the address space cannot name is the larger of the two and
  `prefix_len` is the answer. Stated as a match.
- core/decode/context.rs:377-381 - `RESERVE_CLAMP` becomes usize and the whole
  reservation is computed in usize. A declared size above the cap reserves the
  cap, which is what the u64 `min` already gave.
- cadmpeg/src/inspect/mod.rs:812-820 - the comparison is against the in-memory
  buffer: a start or an end the buffer does not hold, address space included,
  selects the end of the buffer. `saturating_add` becomes `checked_add`.
- iges/parameter.rs:3449 - `census: &Range<u32>` states its own length as an
  index: `census.len()`. No conversion and no saturating subtraction remain.

Non-conversion sites:

- catia/solve/mesh_quotient.rs:3595 - `faces_by_component` is filled by pushing
  a face into a fresh vector, so no component is empty and the `None` arm was
  unreachable. `active_faces` is built by an ascending enumeration, so the face
  that creates an entry is the component's smallest; the map now carries that
  face and the sort key is a usize. Ordering is unchanged: the keys were and
  are the distinct per-component minima.
- catia/solve/mesh_quotient.rs:7300 - the `checked_shl` None is reachable: a
  boundary with 64 or more unresolved uses states more direction choices than
  the work counter holds. The saturation is a real decision and is now spelled
  as a match with the reason, not as a sentinel default.
- inventor/sketch.rs:1806 - the `min` None is reachable: a profile whose
  entities state no source position. The key is now `(first.is_none(), first)`,
  which sorts such a profile after every profile that states a position - the
  order `usize::MAX` gave.

Test: `a_u32_index_widens_to_the_same_value` in
`crates/cadmpeg-core/src/decode/view.rs`. Decisiveness is the census plus the
const assertion: with `usize::BITS >= 32` changed by hand to `>= 128`,
`cargo check -q -p cadmpeg-core` fails with

  error[E0080]: evaluation panicked: a target whose usize is narrower than 32
  bits cannot hold a u32 index
    --> crates/cadmpeg-core/src/decode/view.rs:31:15
  error: could not compile `cadmpeg-core` (lib) due to 1 previous error

Signature changes and their caller inventories (`git grep`):
- `NumericPayload::element_count` / `NumericArray::element_count` u64 -> usize.
  Callers: `creo/src/legacy.rs:120`, `creo/src/decode/build/coverage.rs:910`.
  `meta.rs:182` calls `StringPayload::element_count`, which already returned
  usize and is untouched.
- `RESERVE_CLAMP` u64 -> usize. Only reader: `context.rs:377`.
- `session_ceiling` is new; its five callers are listed above.

+++ crates/cadmpeg-codec-creo/src/legacy.rs
@@ -114,7 +114,7 @@ impl<T> NumericPayload<T> {
     }
 
     /// Number of logical scalar elements represented by this payload.
-    pub fn element_count(&self) -> u64 {
+    pub fn element_count(&self) -> usize {
         match self {
             Self::Scalar { .. } => 1,
             Self::Array(array) => array.element_count(),
